improve subsetting, getting started nb, tests
