+++ tests/test_core/test_core.py
@@ -30,66 +30,6 @@ def _assert_fields_are_view(edata: EHRData):
     assert isinstance(edata.t, ad._core.views.DataFrameView)
 
 
-@pytest.fixture
-def X_32():
-    return np.arange(1, 7).reshape(3, 2)
-
-
-@pytest.fixture
-def X_33():
-    return np.arange(1, 10).reshape(3, 3)
-
-
-@pytest.fixture
-def r_322():
-    return np.array([[[1, 2], [3, 4]], [[5, 6], [7, 8]], [[9, 10], [11, 12]]])
-
-
-@pytest.fixture
-def r_333():
-    return np.arange(1, 28).reshape(3, 3, 3)
-
-
-@pytest.fixture
-def obs_31():
-    return pd.DataFrame({"obs_col_1": [1, 2, 3]}, index=["obs1", "obs2", "obs3"])
-
-
-@pytest.fixture
-def var_21():
-    return pd.DataFrame({"var_col_1": [1, 2]}, index=["var1", "var2"])
-
-
-@pytest.fixture
-def var_31():
-    return pd.DataFrame({"var_col_1": [1, 2, 3]}, index=["var1", "var2", "var3"])
-
-
-@pytest.fixture
-def t_11():
-    return pd.DataFrame({"t_col_1": [1]}, index=["t1"])
-
-
-@pytest.fixture
-def t_21():
-    return pd.DataFrame({"t_col_1": [1, 2]}, index=["t1", "t2"])
-
-
-@pytest.fixture
-def t_31():
-    return pd.DataFrame({"t_col_1": [1, 2, 3]}, index=["t1", "t2", "t3"])
-
-
-@pytest.fixture
-def edata_333(X_33, r_333, obs_31, var_31, t_31):
-    return EHRData(X=X_33, r=r_333, obs=obs_31, var=var_31, t=t_31)
-
-
-@pytest.fixture
-def adata_33(X_33, obs_31, var_31):
-    return ad.AnnData(X=X_33, obs=obs_31, var=var_31)
-
-
 #################################################################
 ### Test combinations of X, r, t during initialization
 #################################################################
@@ -557,11 +497,3 @@ def test_copy_of_obsvar_names(edata_333, adata_33):
     edata_obsvar_subset = edata_obsvar_subset.copy()
     assert not edata_obsvar_subset.is_view
     _shape_test(edata_obsvar_subset, (1, 2, 3))
-
-
-def test_repr():
-    pass
-
-
-# TODO:
-# - allow explictly for r to be only a numpy array or dask array?
